feat: change timestamp to double for sub-second accuracy. This is essential because the CanSat transmits data every 0.5 seconds, so having the timestamp at a 1 second accuracy tends to double the data in 1 second.

+++ Library/Models/DataStamp.cs
@@ -9,7 +9,7 @@
         /// Milliseconds since the Arduino was initialized, registered
         /// when the data was sent.
         /// </summary>
-        public long Timestamp { get; init; }
+        public double Timestamp { get; init; }
 
         /// <summary>
         /// Coordinates registered by the GPS when the data was sent
